# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.AddContainerCheckRequest:
    properties:
      containerID:
        type: integer
      containerType:
        $ref: '#/definitions/schema.ContainerType'
      description:
        type: string
    required:
    - containerID
    - containerType
    - description
    type: object
...
  handlers.PatchContainerCheckRequest:
    properties:
      description:
        type: string
    required:
    - description
    type: object
...
  handlers.UpdateContainerAliasesRequest:
    properties:
      aliases:
        items:
          type: string
        type: array
      containerID:
        type: integer
      containerType:
        $ref: '#/definitions/schema.ContainerType'
    required:
    - containerID
    - containerType
    type: object
  handlers.UpdateFileAliasesRequest:
    properties:
      aliases:
        items:
          type: string
        type: array
      filePath:
        type: string
    required:
    - filePath
    type: object
  models.AliasModel:
    properties:
      alias:
        type: string
      filePath:
        type: string
      id:
        type: integer
      itemId:
        type: integer
      type:
        $ref: '#/definitions/schema.AliasType'
    required:
    - alias
    - id
    - type
    type: object
  models.ContainerCheck:
    properties:
      containerID:
        type: integer
      containerType:
        $ref: '#/definitions/schema.ContainerType'
      description:
        type: string
      id:
        type: integer
    type: object
...
  schema.AliasType:
    enum:
    - epic
    - story
    - task
    - question
    - problem
    - knowledge-node
    - knowledge-bit
    - definition
    - action
    - repetitive-task
    - state
    - file
    type: string
    x-enum-varnames:
    - AliasTypeEpic
    - AliasTypeStory
    - AliasTypeTask
    - AliasTypeQuestion
    - AliasTypeProblem
    - AliasTypeKnowledgeNode
    - AliasTypeKnowledgeBit
    - AliasTypeDefinition
    - AliasTypeAction
    - AliasTypeRepetitiveTask
    - AliasTypeState
    - AliasTypeFile
  schema.ContainerType:
    enum:
    - epic
    - story
    - task
    - question
    - problem
    - knowledge-node
    - knowledge-bit
    - definition
    - action
    - repetitive-task
    - long-task
    - direction
    - state
    type: string
    x-enum-varnames:
    - ContainerTypeEpic
    - ContainerTypeStory
    - ContainerTypeTask
    - ContainerTypeQuestion
    - ContainerTypeProblem
    - ContainerTypeKnowledgeNode
    - ContainerTypeKnowledgeBit
    - ContainerTypeDefinition
    - ContainerTypeAction
    - ContainerTypeRepetitiveTask
    - ContainerTypeLongTask
    - ContainerTypeDirection
    - ContainerTypeState
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Task management dashboard API server.
    **Authentication:** Use `POST /users/login`, copy `accessToken` from the response, then click **Authorize** and paste it as the `access_token` cookie value. Alternatively, login from Swagger on the same host and cookies are sent automatically.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Dashboard API
  version: "1.0"
paths:
...
  /aliases/{alias}:
    get:
      consumes:
      - application/json
      description: Returns an alias by its string
      parameters:
      - description: Alias
        in: path
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AliasModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get alias by string
      tags:
      - aliases
  /aliases/container:
    put:
      consumes:
      - application/json
      description: 'Replaces container aliases: removes absent ones and adds new ones'
      parameters:
      - description: Aliases sync request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateContainerAliasesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AliasModel'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sync aliases for container
      tags:
      - aliases
  /aliases/container/{type}/{id}:
    get:
      consumes:
      - application/json
      description: Returns all aliases linked to the given container type and id
      parameters:
      - description: Container type
        in: path
        name: type
        required: true
        type: string
      - description: Container ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AliasModel'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List aliases for container
      tags:
      - aliases
  /aliases/file:
    put:
      consumes:
      - application/json
      description: 'Replaces file aliases: removes absent ones and adds new ones.
        filePath is the logical relative path under FILES_DIR.'
      parameters:
      - description: Aliases sync request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateFileAliasesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AliasModel'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sync aliases for file
      tags:
      - aliases
  /aliases/file/{filepath}:
    get:
      consumes:
      - application/json
      description: Returns all aliases linked to the given logical relative file path
        under FILES_DIR
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AliasModel'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List aliases for file
      tags:
      - aliases
...
  /container-checks:
    post:
      consumes:
      - application/json
      description: Creates a check entry for the given container
      parameters:
      - description: Check creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddContainerCheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ContainerCheck'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add check to container
      tags:
      - container checks
  /container-checks/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a container check by ID
      parameters:
      - description: Container check ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove container check
      tags:
      - container checks
    patch:
      consumes:
      - application/json
      description: Updates a container check's description
      parameters:
      - description: Container check ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PatchContainerCheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ContainerCheck'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update container check
      tags:
      - container checks
...
schemes:
- http
- https
security:
- AccessTokenCookie: []
securityDefinitions:
  AccessTokenCookie:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"